feat: :sparkles: Punto 6

+++ Classes/Productos.cs
@@ -8,7 +8,7 @@ public class Productos
 {
     public int IdProducto { get; set; }
     public string NombreProducto { get; set; }
-    public double PrecioUnit { get; set; }
+    public int PrecioUnit { get; set; }
     public int Cantidad { get; set; }
     public int StockMin { get; set; }
     public int StockMax { get; set; }
